client: spicec --full-screen=auto-conf do not resize after migration #584318
When running spicec --full-screen=auto-conf and then migrating the machine,
after migration is complete spicec issues a new request to the vdagent to
resize the screen. This is a result of the new "disconnect/connect" migration
introduced recently. There is even functionality to avoid sending a monitor
message twice already, but on disconnect it is explicitly reset. The one line
solution is to not reset _agent_mon_config_sent